Add files via upload
I'm thrilled to share my project with you on GitHub, which I've developed using the Seaborn dataset. I've visualized the data according to my own aspects and insights, and I'm proud of the hard work and dedication that has gone into this project. I've overcome many challenges to complete it, and I'm eager to hear your thoughts and ideas on improving. I'm looking forward to your feedback and guidance as my mentor, and I'm confident that your suggestions will help me refine my skills and knowledge.